refactor: moved trim decorator to the shared module
