+++ packages/core/src/shared/decorators/trim.decorator.ts
